fixed session management issue (#28)
Tons of bizzare behavior related to cookies and session IDs when collecting trace:

When clicking Clear State the session id is being updated if there are any changes to the DB it seesm
When clearing cookies, then visiting products and then orders the session id seems to also change between products/ and orders/ page the first time you do it. Afterwards it persists.
Fix is as follows:

Update the session ID management logic to store session ID locally (so if the cookie is ever deleted we still use the same session ID)
Made every concurrent /generate API call await the same response so that we do not have duplicate sessions / race conditions that may cause session ID to change arbitrarily

Co-authored-by: Sahil Lamba <slamba@Sahils-MacBook-Pro.local>